requirements.txt
altgraph==0.17
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
future==0.18.2
googletrans==2.4.0
idna==2.8
multidict==4.7.5
pefile==2019.4.18
pycparser==2.19
PyInstaller==3.6
python-telegram-bot==12.4.2
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.22.0
six==1.14.0
tornado==6.0.3
urllib3==1.25.8
yarl==1.4.2
